AliM1543C ide: Fix semaphore overload bug
The controller thread is controlled by a semaphore triggered by
other threads. However there wasn't any controller that checked if
the controller thread is ready, which led to the semaphore firing
multiple times at once, crashing the emulator. This fixes the
problem by adding another semaphore that fires when the controller
thread is ready.