segfault seems to be caused by iNumComponents being 8 but acComponents[] only has 7 items. Delete-ing number 8 segfaults, which means that memory wasn't new-ed. During construction of components, they are registered and counted in a counter.
I first tried to change the array to a std::vector, but that left a half constructed item,
failed as well. Adding an unregister function, and calling that in the destructor,
assures that the component that failed to construct (due to disk throwing an error inside its constructor)
is not used later on in other destructors.